Intersection algorithms

Intersections.bk — Function

bk(onmatch::Function, L, findpos::Function=doublingsearch)
bk(L, output=eltype(L[1])[], findpos::Function=doublingsearch)

Computes intersection of posting lists in L using findpos storing the intersection in output.

The method accepting onmatch calls onmatch(L, P) where L are the input lists (in some functions this could be a permutation of the input), and P and array of indices where each match occurred in L[i].

Intersections.baezayates — Function

baezayates(A, B, output=eltype(A)[], findpos=binarysearch)

Computes the intersection between first and second ordered lists using the Baeza-Yates algorithm [cite]. The intersection is stored in output.

Intersections.svs — Function

svs(postinglists, intersect2=baezayates)

Computes the intersection of the ordered lists in postinglists using a small vs small strategy. Accepts an intersection algorithm of two sets.

Intersections.imerge2 — Function

imerge2(A, B, output=eltype(A)[])

Computes the intersection of A and B (sorted arrays), using a merge-like algorithm, and stores it in output.

Union algorithms

Intersections.umerge — Function

umerge(L, output=eltype(L[1])[], P=nothing; t=1)
umerge(onmatch::Function, L, P=nothing; t::Int=1)

Merges posting lists in L and saves the union in output. The method accepts a callback function onmatch that is called whenever an object occurs in at least t posting lists.

Arguments:

L: The array of posting lists, the array can be destroyed in the process.

P: The array of current positions in posting lists, i.e., initial state as an array of ones of size $|L|$.

Keyword arguments

t: A positive treshold (number of occurrences) to push or apply the callback function, i.e., $t=1$ means for union.

About the callback function

The callback signature is onmatch(L, P, m) where:

L is an array of postings lists, the same of the input but maybe sorted and perhaps some lists could be removed by the process.

P a list of indices of the current position of L, e.g., L[i][P[i]] to obtain the current position of i-th list.

m the number of occurences of the current element pointed by P. Note that all L[i][P[i]] are the same for 1 ≤ i ≤ m (it is an alignment position).

Searching algorithms

Intersections.binarysearch — Function

binarysearch(A, x, sp=1, ep=length(A))

Finds the insertion position of x in A in the range sp:ep

Intersections.doublingsearch — Function

doublingsearch(A, x, sp=1, ep=length(A))

Finds the insertion position of x in A, starting at sp

Intersections.doublingsearchrev — Function

doublingsearchrev(A, x, sp=1, ep=length(A))

Finds the insertion position of x in A, starting at the end

Intersections.seqsearch — Function

searchrev(A, x, sp=1, ep=length(A))

Sequential search, i.e., it starts from sp to ep

Intersections.seqsearchrev — Function

seqsearchrev(A, x, sp=1, ep=length(A))

Reverse sequential search, i.e., it starts from ep to sp